fix #109 issue. Remove double decoding on server side. Fix link to raw object. (#113)
* fix #109 issue. There is two parts of contribution.
Firs: remove double decoding on server side for all parameters for CatalogObject Controllers for all REST API requests, as Spring already decode all parameters. Double decoding was causing the user-not expecting transformation of ObjectName parameter.
Second: the link to raw object provided in response metadata was not correct for special symbols.
The double encoding inside link was solved. The space is correctly encoded in link. So the raw object can be correctly retrieved by specified link

This behavior is checked with a created specific test.