Improve performance of job-planner filtering
When the association status filter is anything but UNPLANNED, we can query the job-planner service for all associated objects and retrieve from the db catalog object entities according to this list.

In case of UNPLANNED, it is not possible to do this filtering and the performance will not be improved.

Also, add pa.catalog.db.items.max.size to control the SQL in expression size